Boolean creates issues at the SOAP level.
With the latest change, clear_datastore was sent on the wire as a
FalseClass (or TrueClass) and we found this error:

FATAL: [set_parameters] We saw an unexpected error of the type SOAP::FaultError with the following message: can't convert false into String, with trace: failed to serialize detail object: Cannot map SOAP::Mapping::SOAPException to SOAP/OM.

The error comes from the appcontroller_client make_call, and it appears on
the client side (in this case the head node trying to pass parameters to
the others nodes). The error happened on clear_datastore since it's the
ony one that seems to have been translated into false (not the string
'false'), while the other TrueClass are passed as String. This patch force
all the options TrueClass to be String, and we check them manually as
String. When switching to rest we should be able to convert back into
booleans.