Changed the tag logic for bootstrap.
First of all, one cannot specify --branch and --tag at the same time in
the bootstrap script. We also added a 'dev' tag that allows to pull from
the HEAD.

For upgrade, we don't want to change branch: that will need to
be done by hand before calling the script. The idea is to disable
--branch when upgrading, then if the current branch is master, the upgrade
will be the specified tag. If the current branch is not master, we simply
do a pull.