Use a longer hysterisis cycle for VMs.
The thresholds we use to scale up/down on single nodes, needs to be longer
when applied to instances, since the time to boot them is loonger, and we
may want to leave one instance running longer to remove that start up time
if we can. Don't go over 1/2 hour since some cloud platform charge by the
hour.