Fix reboot issue.
When we stop appscale, we need to be sure we remove some of the old state.
We should use monit standard configuration in place, and let monit restart
at boot: if appscale was shutdown properly, no *cfg are present, so
nothing will start, if the instance was accidentally rebooted, appscale
will come up via monit.