Upon shutdown, make login node to contact all nodes.
In case the other nodes are not visible by the tools, the login node
should terminate appscale on all other nodes in the deployments. We do it
in a thread, to ensure we don't stop cleaning the node.

Make terminate.rb to properly restart monit when removing the
configuration. We should also remove only appscale configuration not all
od them.