Use ZK as backend storage for deployment state.
This patch will make zookeeper (instead of the local disk) the persistent
storage for the deployment state. With this change, each AC now reads
routinely the state from ZK: thus we changed the way appservers are
started to make sure the state is propagated to the head node.

The system is now able to better tolerate restarts and failures on nodes:
the main requirement is that the zookeeper cluster is up and running.