Set num_tokens to 1
This is for compatibility with existing deployments. Cassandra's
behavior changed in 2.1.15. Firstly, it actually uses this value
if initial_token is set to an empty string. Secondly, "Specifying
initial_token will override this setting on the node's initial
start, on subsequent starts, this setting will apply even if
initial token is set."