Abandon atomicity for non-transactional batches
This commit uses a batch per entity mutation instead of a single
Cassandra batch for the entire datastore batch as long as it is
not part of a transaction. This is to minimize large Cassandra
batches as they can have a large impact on performance and stability.
According to the Cloud Datastore documentation: "Writing multiple
entities in a single operation does not guarantee that the writes will
take place atomically unless the operation is performed inside a
transaction."