Check the return value from done_uploading
done_uploading adds the node to the list of 'app-hosters' in zookeeper.
This is used to find which nodes has a recent copy of the application when
it's time to refresh the local copy. It could happen that zookeeper is not
available at that time, so we need to retry few times and print an error
message if something goes wrong.

Reference:
https://ocd.appscale.com:8080/job/Blobstore/1149/artifact/appscale-logs/192.168.103.178/appscale/controller-17443.log