Make sure we check the return from UAS
The UserAppServer calls to get_all_apps, and get_all_users may now raise
an exception. Also ensure that each time a reserved app's appserver is
started a fresh copy of the code is pulled from the shadow. This is very
important for the dashboard in particular upon restart, since there may be
a race condition with the headnode (which re-package the dashboard code),
so having the appengine node to pull the code as late as possible ensure
the freshness of the deployment secret.