Fix logic in stopping application.
The previous logic used the list of application coming from the
UserAppServer, then add the application we have in /var/apps, then go
through them to see which one is stopped, and stop it.  This logic is
too wasteful, since we keep trying to stop application stopped long time
ago.

The new logic check what we have in /var/apps (that is application we
may be running) and check with UserAppServer and with the shadow node
state (@app_info_map) if this app is still running or not. Since the
clanup part remove the app in /var/apps, this process can only be done
once per application removal.