Give the Java AppServer a max heap size
The max heap size is set to 140MB less than the app's max memory.
In practice, the process tends to use about 120MB more (according
to monit) than the max heap size given to Java.

Previously, the default heap size was used, which was normally
~1GB. This change should prevent monit from killing the Java
AppServer processes. Instead, the app will encounter an
OutOfMemoryError if it tries to allocate more than the max heap
size allows.