Removed num_appengine instance variables
Use directly @options[appengine] instead of a local variables
num_appengines. This allows the set_property to work directly on a
property coming from AppScalefile (that is the user will know what to
change).

Also we enabled the autoscale options. When set to false, the system
will not scale at all appservers.