Removed detection and change of private IP
The code to detect private IP is there, but we don't currently have a
way to recover from a change of private IP. In this situation we should
notify the master, ask it to change our IPs, etc ... It will need to
wait when we store @nodes in zookeeper and the headnode can detects
changes on slaves.

Don't use ifconfig to detect private IPs by Socket.