Serialize access to the AppManagerServer
Make sure that update() will not call the AppManagerServer directly on
all nodes: instead notify all AppControllers that an application restart
is needed.  Also use AMS_LOCK when calling the AppMangerServer to stop
an application.  With this serialization of calls, we re-enabled the
thread in the AppManagerServer to wait for the app.