Minimize the time a terminate appserver can be UP
The logic to gently terminate the AppServer, is to keep track of the
to-be-terminated AppServer, and set HAProxy to drain them. This
flag is transient and a reload of HAProxy will reset it. This PR,
minimize the number of HAProxy reloads, checking the new and old
configuration (and avoiding a reload if not needed), then make sure we
set the drain for each operation that could have HAProxy reloaded.