Nodes will adjust roles each duty cycle
This commit, adds the ability for the nodes to adjusts their role each
duty cycle. Upon reloading the appcontroller state from the head node,
each node will check if the role list changed, and if so, it will
re-invoke start_stop_api_service (modified by this commit) to handle the
starting/stopping or new/stopped roles.