Decrease gc_grace_seconds on ephemeral data
The major downside of lowering this value is that it's possible
for data to be (unintentionally) undeleted if a node goes offline
and the tombstones were not replicated.

For ephemeral tables like transactions and pull_queue_leases
(used for stats), this downside is not relevant since all data
is inserted with a TTL.

For pull_queue_tasks_index, undeleted index entries will not
affect correctness, and it will barely affect performance.

The major benefit is reduced disk space usage (for the transactions
table in particular) and increased performance for pull queue
index lookups since tombstones will be deleted earlier.