Remove unused modification to labs
This MapReduce library is no longer part of the labs jar. I'm not
sure if it ever was. It was not present even in the 1.4.0 SDK.
We don't have the hadoop jar that it references, so I think
it's safe to say that any app that uses this library won't work
anyway on our current release.