Kill instances when then exceeded memory limit
Previously, Monit would just try to teminate it gently, allowing
the instance to finish serving open requests. In some cases, this
limit was not long enough (30 seconds). In other cases, it was too
long. It triggered the controller to try to stop the instance
through the AppManger, resulting in a sequence that caused monit
to restart the instance after the AppManger unmonitored it.