Clean up group locks when possible
This allows the datastore server to clean up entity group locks
when it can. When a failure happens during a normal batch or
before a large batch has been marked as "applied", the server can
release the locks before returning the error.

This also makes the server more persistent after it passes the
point of no return for large batches. The additional retries
makes it less likely for the server to give up and leave the
locks around for the transaction groomer to clean up.