Handle reads with the index past the last block
Return an empty string when asked for a chunk that's outside of
the stored blob.

If the end of a blob happens to match the end of a block, a
BlobReader may ask for the next chunk as it does not know it has
read the end of the blob. In that case, it does not make sense to
claim that the blob does not exist. Rather, like a typical file
handle, it should return an empty byte string.