Wait for all hosts before creating tables
The most common error we see when running the prime script is
that one or more hosts are immediately marked as "down" after
the connect call finishes. Those machines later rejoin and cause
a schema disagreement. This change is meant to prevent the script
from continuing immediately if not all hosts are available.