Remove indexless sorted ancestor queries
In GAE, sorted ancestor queries require a composite index. It's
not really possible to order the results without fetching the
entire entity group, which is not reasonable for large entity
groups. It's also not possible to use a cursor since the order
of the index does not match the order of the results.