Use different error when changing a deleted task
In GAE, extending the lease of a task that doesn't exist returns
an error indicating that the lease has expired. This is a
different behavior than when a task has been tombstoned. Since we
do not tombstone pull queue tasks in the Cassandra implementation,
we treat recently deleted tasks the same as tasks that have never
been created.