Restart instances on login property update
This moves the burden of enforcing the correct login property
value to the InstanceManager. Without this, there is a chance
the InstanceManager may not restart the instances on a property
update. For example, if the InstanceManager is offline when
the controller clears and then repopulates the assignments list,
the InstanceManager will consider the assignments to be fulfilled
upon restart.