Use a patched version of Cassandra
This uses a version of Cassandra that was built from source. The
patch from https://issues.apache.org/jira/browse/CASSANDRA-12942
is the only modification.

This fixes an issue with `nodetool status` and `nodetool ring`
occasionally not working. The rebalance script, Cassandra
monitoring, and service start all rely on those tools.