Add new method of setting instance configuration
Previously, the dev_appserver would pass instance configuration
via stdin and read the port that the instance decided to use
via stdout. The new method writes the configuration to a file,
and the instance writes its port to a separate file. Both
temporary files are cleaned up shortly after creation. The
advantage of the new method is that instances can write to stdout
without affecting what the dispatcher uses for the instance port.

Currently, this is just used for the python27 runtime.

This change was introduced in the 1.8.3 SDK.