Wrap unrecognized API errors in RuntimeError
For python-based runtimes, this allows the runtime instance to
raise a normal exception when the API Server encounters an error
that isn't clearly defined.

This also logs a stack trace when the API Server suffers an
uncaught exception, making it easier to determine the source of
an issue during an API call.

This change was introduced in the 1.8.4 SDK.