Terminate not started AppServers first
When the autoscaler decides it is time to downscale, the previous logic
would pick AppServers from the newly created machines, in order to
facilitate the removal of instances if not used. We should start
removing first AppServers not started yet, in order to leave a minimum
number of *running* AppServers.