Removed useless Mutex
The mutex was introduced with the idea of having the possibility of
multiple update_node_info_cache running at the same time. Since the
design is now chnaged, and there is only one invocation and the spawn
threads are joined within the function, there is no need for this any
longer.