Run datastore servers in a custom slice
The memory cgroup is not as stable as the namespace cgroup that
the process is running in. It's possible for systemd to reset the
process's memory cgroup to the parent slice's memory cgroup.

This change keeps track of datastore processes in an
"appscale-datastore" slice to ensure that systemd won't move them.