Remove caching of local IP
HelperFunction used a class variable to cache of the local IP used. It
was set by djinn upon start (in find_me_in_locations) and it was used
only sparingly with few services (by djiin to start zookeeper, and the
datastore connector). Since only few services were using it, this PR
removes it, and default to look for the default local IP of the node.