Removed cached local IP
HelperFunctions was using a class variable to track the local IP, but it
was not quite a cache, since it was set by djinn (or by calling
local_ip). This variable was used sparingly (starting zookeeper and for
the datastore): since not all services were using it, this could in
theory cause some inconsistencies on multi-networked systems. This PR
removes the inconsistency and print warnings if djinn is using a local
IP that is not the default one.