Do not override hostname
Use whatever hostname the machine uses and let rabbitmq to use it. To
properly have rabbitmq clustering, the hostname (not fqdn just name) of
the machines needs to be resolvable from all TQ machines. In case the
DNS is not setup correctly, /etc/hosts can be modifies on all machines
to achieve the same.