Limit work in set_parameters
set_parameters (called by the tools during up) was unlocking the main
threads and then proceeded to do more checks on the configuration. This
commit, limit the work done in set_parameters delegating some checks to
enforce_options and ensuring that we unlock the main thread only if both
options and layout have been validated.