Handle scheme redirects with the dispatcher
This is less efficient than giving nginx the burden of handling
client redirects. However, it makes it much simpler to implement
the dispatch configuration, and the instance dispatcher can later
be optimized if necessary.

Since the dispatcher is already handling static files (usually)
and login redirects, it makes sense for it to handle scheme
redirects as well.